Plex/requirements.txt
alive-progress
PlexAPI
tmdbapis
tmdbv3api
python-dotenv
tvdb_v4_official
ruamel.yaml
requests
pathvalidate
imdbpy
python-magic